fix: sidebar component
